% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbtcomp.R
\name{bbtcomp}
\alias{bbtcomp}
\title{Generates a BBT model from a table with the results of algorithms on
different data sets}
\usage{
bbtcomp(
  tabx,
  tabsd = NULL,
  dbcol = 1,
  lrope = NULL,
  lrope_value = 0.4,
  paired = TRUE,
  deal_with_ties = c("spread", "forget", "davidson", "add", "random"),
  hyper_prior = NULL,
  scale = NULL,
  ...
)
}
\arguments{
\item{tabx}{An array or data frame, where lines are data sets and columns are algorithms.
There are three basic use case.
* only \code{tabx}, and there is a column that contains the name or an id for each data set. The dbcol
parameter indicates this column. In this case, there can be multiple lines for the same
data set, where the lines are the results on the different folds of the data set. In this case
the computation of wins and losses can use the local ROPE concept. If the \code{lrope} parameter is true
then local ROPE will be used. Also, in this case, the algorithm can use the paired version
of the local ROPE (if all folds for this data set are the same for all algorithms). If the
\code{paired} parameter is set, the algorithm will use the paired local ROPE
* \code{tabx} and \code{tabsd} are given. In this case, local ROPE can be used (if \code{lrope} is TRUE) but not
the paired version (\code{paired} is silently ignored). If \code{lrope}
is FALSE, \code{tabsd} is silently ignored.
* only \code{tabx} is given and there is no data set column (\code{dbcol} is set to 0). In this case
local ROPE cannot be used (and \code{lrope} and \code{paired} are silently ignored.)}

\item{tabsd}{A data frame or an array with the standard deviation of the fold measures for each data set and
algorithm. If given,  the parameter \code{lrope} determines whether local ROPE will or will
not be used to compute wins and losses,}

\item{dbcol}{Column number or name were the data set indicator is located, or 0 for no such column}

\item{lrope}{Whether to use local ROPE in the pairwise comparisons}

\item{lrope_value}{The value of the local ROPE threshold (default 0.4)}

\item{paired}{Whether to use paired local ROPE calculations}

\item{deal_with_ties}{How to deal with ties in the wintable. The options are:
- spread - half (rounded up) of each ties is added to each algorithm.
- forget: remove the ties column.
- add: add the ties to each algorithm.
- random: add each tie randomly to one of the algorithms.
- davidson: do nothing with the ties. TO be used with the Davidson Bayesian model.}

\item{hyper_prior}{The code of the hyper-prior for the sigma

\if{html}{\out{<div class="sourceCode">}}\preformatted{* 0 = lognormal(0.5) - the default,
* 1 = lognormal(scale),
* 2 = cauchy(scale),
* 3 = normal(scale)
}\if{html}{\out{</div>}}}

\item{scale}{The scale of the hyper prior for the sigma parameter}

\item{...}{Extra parameters to be passed to the MCMC algorithm (cmdstanr)}
}
\value{
A BBT model. The model contains:

\if{html}{\out{<div class="sourceCode">}}\preformatted{   - `model`: a fit object from cmdstanr
   - `wintable`: the wintable generated from the data
   - `davidson` whether the Davidson model was used
}\if{html}{\out{</div>}}
}
\description{
Generates a BBT model from a table with the results of algorithms on
different data sets
}
\examples{
\donttest{
ss <- ll[1:80, 1:6] # first 20 data sets and 5 algorithms

options(mc.cores = parallel::detectCores(logical = FALSE))

m1 <- bbtcomp(ss)
m2 <- bbtcomp(ss, lrope = FALSE, iter_sampling = 2000, seed = 1234)
m3 <- bbtcomp(ss, lrope_value = 0.2, paired = FALSE)
m4 <- bbtcomp(ss, deal_with_ties = "davidson", hyper_prior = 3, scale = 3.0)

}
}
