% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_loo.R
\name{get_loo}
\alias{get_loo}
\title{Compute the loo for the model}
\usage{
get_loo(mod)
}
\arguments{
\item{mod}{A BBT model}
}
\value{
The loo, as computed by the loo package
}
\description{
Compute the loo for the model
}
\examples{
\donttest{
ss <- ll[1:80, 1:6] # first 20 datasets and 5 algorithms

m1 <- bbtcomp(ss)
get_loo(m1)
}
}
