% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_loo.R
\name{get_waic}
\alias{get_waic}
\title{Compute the WAIC for the model}
\usage{
get_waic(mod)
}
\arguments{
\item{mod}{A BBT model}
}
\value{
the waic, as computed by the loo package
}
\description{
Compute the WAIC for the model
}
\examples{
\donttest{
ss <- ll[1:80, 1:6] # first 20 data sets and 5 algorithms

m1 <- bbtcomp(ss)
get_waic(m1)
}
}
