% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ll}
\alias{ll}
\title{Accuracy of 16 classifiers on 132 data sets}
\format{
A data set with 17 columns and 526 rows:
\describe{
\item{db}{The name of the data base}
\item{lgbm}{ Results for the lgbm classifier}
...
}
}
\usage{
ll
}
\description{
The accuracy of 16 different classifiers on 132 datasets
as computed by the author
}
\keyword{datasets}
