% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wintable.R
\name{make_wintable}
\alias{make_wintable}
\title{Construct a wintable from a table of fold results for each combination of algorithm and data set}
\usage{
make_wintable(
  tabx,
  tabsd = NULL,
  dbcol = 1,
  lrope = NULL,
  lrope_value = 0.4,
  paired = TRUE,
  deal_with_ties = c("spread", "forget", "davidson", "add", "random")
)
}
\arguments{
\item{tabx}{An array or data frame, where lines are data sets and columns are algorithms.
There are three basic use case.
* only tabx, and there is a column that contains the name or an id for each data set. The dbcol
parameter indicates this column. In this case, there can be multiple lines for the same
data set, where the lines are the results on the different folds of the data set. In this case
the computation of wins and losses can use the local ROPE concept. If the \code{lrope} parameter is true
then local ROPE will be used. Also, in this case, the algorithm can use the paired version
of the local ROPE (if all folds for this data set are the same for all algorithms). If the
\code{paired} parameter is set, the algorithm will use the paired local ROPE
* tabx and tabsd are given. In this case, local ROPE can be used (if \code{lrope} is TRUE) but not
the pairde version (\code{paired} is silently ignored). If \code{lrope}
is FALSE, \code{tabsd} is silently ignored.
* only tabx is given and there is no data set column (\code{dbcol} is set ot 0). In this case
local ROPE cannot be used (and \code{lrope} and \code{paired} are silently ignored.)}

\item{tabsd}{A data frame or an array with the standard deviation of the fold measures for each data set and
algorithm. If given,  the parameter \code{lrope} determines whether local ROPE will or will
not be used to compute wins and losses,}

\item{dbcol}{Column number or name were the data set indicator is located, or 0 for no such column}

\item{lrope}{Whether to use local ROPE in the pairwise comparisons}

\item{lrope_value}{The value of the local ROPE threshold (default 0.4)}

\item{paired}{Whether to use paired local ROPE calculations}

\item{deal_with_ties}{How to deal with ties in the wintable. The options are:
- spread - half (rounded up) of each ties is added to each algorithm.
- forget: remove the ties column.
- add: add the ties to each algorithm.
- random: add each tie randomly to one of the algorithms.
- davidson: do nothing with the ties. TO be used with the Davidson Bayesian model.}
}
\value{
A wintable object
}
\description{
Construct a wintable from a table of fold results for each combination of algorithm and data set
}
\examples{
\donttest{
w1 <- make_wintable(ll)

w2 <- make_wintable(as.matrix(ll[, -1]), dbcol = 0)

w3 <- make_wintable(ll, lrope_value = 0.3, paired = TRUE)

w4 <- make_wintable(ll, deal_with_ties = "random", paired = FALSE)

}
}
