% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppc}
\alias{ppc}
\title{Plot the Posterior Predictive Check}
\usage{
ppc(modout, maxshow = 20)
}
\arguments{
\item{modout}{A BBT model}

\item{maxshow}{The maximum number of plots to display}
}
\value{
A ggplot2 plot
}
\description{
Plot the Posterior Predictive Check
}
\examples{
\donttest{
m1 = bbtcomp(ll, lrope = FALSE, deal_with_ties = "davidson")
ppc(m1)
ppc(m1, maxshow = 30)
}
}
