% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_summary.R
\name{ppc_summary}
\alias{ppc_summary}
\title{Summary of the PPC for all parameters}
\usage{
ppc_summary(modout)
}
\arguments{
\item{modout}{A BBT model}
}
\value{
A data.frame with the different HDI and the proportion of data that
falls within the HDI
}
\description{
Summary of the PPC for all parameters
}
\examples{
\donttest{
m1 = bbtcomp(ll, lrope = FALSE, deal_with_ties = "davidson")
m2 = bbtcomp(ll,  deal_with_ties = "spread")
ppc_summary(m1)
ppc_summary(m2)
}
}
