% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_wintable.R
\name{print_wintable}
\alias{print_wintable}
\title{Prints a wintable}
\usage{
print_wintable(tablex)
}
\arguments{
\item{tablex}{A wintable}
}
\value{
A data.frame with the wintable
}
\description{
Prints a wintable
}
\examples{
\donttest{
w1 <- make_wintable(ll)
print_wintable(w1)

w2 <- make_wintable(ll, lrope = FALSE)
print_wintable(w2)

w3 <- make_wintable(ll, lrope_value = 0.3)
print_wintable(w3)

w4 <- make_wintable(ll, deal_with_ties = "random", paired = FALSE)
print_wintable(w4)

}
}
