% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ppc.R
\name{table_ppc}
\alias{table_ppc}
\title{Summary of the PPC for all parameters}
\usage{
table_ppc(modout)
}
\arguments{
\item{modout}{A BBT model}
}
\value{
A data.frame with the different HDI and the proportion of data that
falls within the HDI
}
\description{
Summary of the PPC for all parameters
}
\examples{
\donttest{
m1 = bbtcomp(ll, lrope = FALSE, deal_with_ties = "davidson")
m2 = bbtcomp(ll,  deal_with_ties = "spread")
table_ppc(m1)
table_ppc(m2)
}
}
