% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pwin.R
\name{table_pwin}
\alias{table_pwin}
\title{Print a summary of the probabilities of an algorithms being better than another}
\usage{
table_pwin(
  modout,
  selected = NULL,
  control = NULL,
  short = TRUE,
  rope = c(0.45, 0.55),
  columns = c("median", "mean", "low", "high", "delta", "above.50", "in.rope"),
  hdi = 0.89,
  ndigits = 2
)
}
\arguments{
\item{modout}{A BBT model}

\item{selected}{List of algorithms names to be printed. NULL for all}

\item{control}{An algorithm name, to be seen as the control. Only the
probabilities of the other algorithms against the control will be
printed}

\item{short}{Print only the short summary (mean, delta, above.50 and in.rope)}

\item{rope}{The low and the high values of ROPE}

\item{columns}{Which columns to print in the summary}

\item{hdi}{HDI for the low/high  and delta  columns in the summary}

\item{ndigits}{Number of digits to print the probabilities}
}
\value{
A data.frame
}
\description{
Print a summary of the probabilities of an algorithms being better than another
}
\examples{
\donttest{
m1 = bbtcomp(ll)
table_pwin(m1, control = "rf", columns = c("median","delta","in.rope"))
table_pwin(m1, selected = c("svm","lda","gbm","passive"), rope = c(0.48, 0.52))
table_pwin(m1, short = FALSE, hdi = 0.95)
table_pwin(m1, columns = c("mean","low", "high"), hdi = 0.90)
}
}
