% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_wintable.R
\name{table_wintable}
\alias{table_wintable}
\title{Prints a wintable}
\usage{
table_wintable(tablex, which = "pos")
}
\arguments{
\item{tablex}{A wintable}

\item{which}{Whether to print the wintable after the processing of ties ("pos")
or before the processing ("pre") or both  ("both")}
}
\value{
A data.frame with the wintable
}
\description{
Prints a wintable
}
\examples{
\donttest{
w1 <- make_wintable(ll)
table_wintable(w1)

w2 <- make_wintable(ll, lrope = FALSE)
table_wintable(w2)

w3 <- make_wintable(ll, lrope_value = 0.3)
table_wintable(w3)

w4 <- make_wintable(ll, deal_with_ties = "random", paired = FALSE)
table_wintable(w4)

}
}
